Chapter-04/requirements.txt
email_validator==2.2.0
fastapi==0.115.11
httpx==0.28.1
passlib==1.7.4
pydantic==2.10.6
pyotp==2.9.0
python_jose==3.4.0
SQLAlchemy==2.0.38
